+++ lib/circogen/circularinit.c
@@ -64,7 +64,7 @@ void circo_init_graph(graph_t * g)
 {
     setEdgeType (g, EDGETYPE_LINE);
     /* GD_ndim(g) = late_int(g,agfindattr(g,"dim"),2,2); */
-    Ndim = GD_ndim(g) = 2;	/* The algorithm only makes sense in 2D */
+    Ndim = GD_ndim(agroot(g)) = 2;	/* The algorithm only makes sense in 2D */
     circular_init_node_edge(g);
 }
 
@@ -273,6 +273,4 @@ void circo_cleanup(graph_t * g)
 	gv_cleanup_node(n);
     }
     free(GD_neato_nlist(g));
-    if (g != agroot(g)) 
-	agclean (g,AGRAPH,"Agraphinfo_t");
 }