+++ lib/fdpgen/fdpinit.c
@@ -125,8 +125,6 @@ static void cleanup_subgs(graph_t * g)
 	cleanup_subgs(subg);
     }
     free (GD_clust(g));
-    if (g != agroot(g))
-	agdelrec(g, "Agraphinfo_t");				
 }
 
 static void fdp_cleanup_graph(graph_t * g)