+++ lib/patchwork/patchworkinit.c
@@ -152,8 +152,7 @@ void patchwork_layout(Agraph_t *g)
 static void patchwork_cleanup_graph(graph_t * g)
 {
     free(GD_neato_nlist(g));
-    if (g != agroot(g))
-        agclean(g, AGRAPH , "Agraphinfo_t");
+    free(GD_clust(g));
 }
 
 void patchwork_cleanup(graph_t * g)