+++ CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - missing Perl includes patch #2262
 - smyrna: incorrect tokenization in frmobjectui.c:set_attr_object_type #2259
 - [Dot] Record shape+UTF+" | "=Eats spaces. #925
+- Memory leak in osage
+- Segmentation fault when running test example neatopack.c #1800
+- Memory leak in dot when using clusters
+- Memory leak in patchwork when using clusters
+- Subgraph layout and rendering
 
 ## [5.0.0] – 2022-07-07